requirements.txt
python-telegram-bot==20.0
rembg
pillow
opencv-python-headless
numpy
